django_admin/requirements.txt
Django==4.2.16
django-environ==0.11.2
psycopg2-binary==2.9.9
pytest==7.4.3
pytest-django==4.8.0
pytest-asyncio==0.21.1
factory-boy==3.3.0
pytest-mock==3.12.0
requests-mock==1.11.0
coverage==7.4.0